[refact] refactoring file with new dir structure
move files & update meson
- deepseek
- ernine

**Self evaluation:**
1. Build test:	 [X]Passed [ ]Failed [ ]Skipped
2. Run test:	 [X]Passed [ ]Failed [ ]Skipped
