+++ Applications/CausalLM/models/meson.build
@@ -14,3 +14,6 @@ subdir('qwen3_moe')
 subdir('qwen3_slim_moe')
 subdir('qwen3_cached_slim_moe')
 subdir('gemma3')
+subdir('ernie4_5')
+subdir('deepseek_v2_lite')
+subdir('deepseek_v2_lite_cached_slim')
